Added more tests and updated tests #44
