--- tests/integration/integration_test.yml
@@ -1,26 +1,0 @@
----
-- hosts: localhost
-  tasks:
-    - name: "Integration testing repository information module"
-      script: ./targets/collaborator_information/{{ item }}.py TOKEN ORGANIZATION_NAME
-      loop:
-        - collaborator_information_regular
-      args:
-        executable: Python3
-      register: result
-
-    - assert:
-        that:
-          result is changed
-
-    - name: "Integration testing repository information module enterprise"
-      script: ./targets/collaborator_information/{{ item }}.py TOKEN ORGANIZATION_NAME ENTERPRISE_URL
-      loop:
-        - collaborator_information_enterprise
-      args:
-        executable: Python3
-      register: result
-
-    - assert:
-        that:
-          result is changed
