+++ tests/integration/targets/collaborator_information/scripts/collaborator_information_enterprise.py
@@ -70,102 +70,146 @@ class TestCollaboratorRepositoryModule(unittest.TestCase):
     API_KEY = 'token'
     ORGANIZATION_NAME = 'Organization Name'
     ENTERPRISE_URL = None
+    USER = None
 
-    # def test_pass_list_collaborators_enterprise_github(self):
-    #     g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
-    #     repo_list = ["testing-repo-private", "testing-repo-internal", "testing-repo-public"]
+    def test_pass_list_collaborators_enterprise_github(self):
+        g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
+        repo_list = ["testing-repo-private", "testing-repo-internal", "testing-repo-public"]
 
-    #     if len(repo_list):
-    #         for i in range(len(repo_list)):
-    #             repo_list[i] = self.ORGANIZATION_NAME + "/" + repo_list[i]
+        if len(repo_list):
+            for i in range(len(repo_list)):
+                repo_list[i] = self.ORGANIZATION_NAME + "/" + repo_list[i]
 
-    #     output = get_collaborators(g, repo_list)
-    #     collaborator_list = list()
-    #     for collaborator in output[repo_list[0]]:
-    #         collaborator_list.append(collaborator['login'])
+        output = get_collaborators(g, repo_list)
+        collaborator_list = list()
+        for collaborator in output[repo_list[0]]:
+            collaborator_list.append(collaborator['login'])
 
-    #     assert 'je652917' in collaborator_list
+        assert self.USER in collaborator_list
 
-    # def test_pass_check_permissions_enterprise_github(self):
-    #     g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
-    #     repo_list = ["testing-repo-private", "testing-repo-internal", "testing-repo-public"]
+    def test_pass_check_permissions_enterprise_github(self):
+        g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
+        repo_list = ["testing-repo-internal"]
 
-    #     if len(repo_list):
-    #         for i in range(len(repo_list)):
-    #             repo_list[i] = self.ORGANIZATION_NAME + "/" + repo_list[i]
+        if len(repo_list):
+            for i in range(len(repo_list)):
+                repo_list[i] = self.ORGANIZATION_NAME + "/" + repo_list[i]
 
-    #     output = get_collaborators(g, repo_list)
+        output = get_collaborators(g, repo_list)
 
-    #     collaborator_list = list()
-    #     for collaborator in output[repo_list[0]]:
-    #         collaborator_list.append(collaborator['login'])
+        collaborator_list = list()
+        for collaborator in output[repo_list[0]]:
+            collaborator_list.append(collaborator['login'])
 
-    #     for i in range(0, len(collaborator_list)):
-    #         if output[repo_list[0]][i]['login'] == 'je652917':
-    #             collaborator_permission = {'permissions': output[repo_list[0]][i]['permissions']}
+        for i in range(0, len(collaborator_list)):
+            if output[repo_list[0]][i]['login'] == self.USER:
+                collaborator_permission = {'permissions': output[repo_list[0]][i]['permissions']}
 
-    #     assert 'admin=False' in str(collaborator_permission['permissions'])
+        assert 'admin=False' in str(collaborator_permission['permissions'])
+        assert 'pull=True' in str(collaborator_permission['permissions'])
+        assert 'push=False' in str(collaborator_permission['permissions'])
 
-    # def test_pass_add_and_delete_collaborator_to_enterprise_github_repo(self):
-    #     g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
-    #     target_repos = ["testing-repo-internal"]
-    #     collaborators_to_add = {"nk479217":"admin"}
-    #     collaborators_to_remove = ["nk479217"]
+    def test_pass_add_and_delete_collaborator_to_enterprise_github_repo(self):
+        g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
+        target_repos = ["testing-repo-public"]
+        collaborators_to_add = {self.USER:"pull"}
+        collaborators_to_remove = [self.USER]
 
-    #     collaborator_added = False
-    #     collaborator_deleted = False
+        collaborator_added = False
+        collaborator_deleted = False
 
-    #     #needed from ansible (list of dict [name, permission])
-    #     if(len(collaborators_to_add) > 0):
-    #         if(len(target_repos) > 0):# needed from ansible (list)
-    #             add_collaborators(g, target_repos, collaborators_to_add, self.ORGANIZATION_NAME)
+        #needed from ansible (list of dict [name, permission])
+        if(len(collaborators_to_add) > 0):
+            if(len(target_repos) > 0):# needed from ansible (list)
+                add_collaborators(g, target_repos, collaborators_to_add, self.ORGANIZATION_NAME)
 
-    #     target_repos_copy = target_repos.copy()
-    #     if len(target_repos_copy):
-    #         for i in range(len(target_repos_copy)):
-    #             target_repos_copy[i] = self.ORGANIZATION_NAME + "/" + target_repos_copy[i]
+        target_repos_copy = target_repos.copy()
+        if len(target_repos_copy):
+            for i in range(len(target_repos_copy)):
+                target_repos_copy[i] = self.ORGANIZATION_NAME + "/" + target_repos_copy[i]
 
-    #     output = get_collaborators(g, target_repos_copy)
-    #     collaborator_list = list()
-    #     for collaborator in output[target_repos_copy[0]]:
-    #         collaborator_list.append(collaborator['login'])
+        output = get_collaborators(g, target_repos_copy)
+        collaborator_list = list()
+        for collaborator in output[target_repos_copy[0]]:
+            collaborator_list.append(collaborator['login'])
 
-    #     assert 'nk479217' in collaborator_list
+        assert self.USER in collaborator_list
 
-    #     if(len(collaborators_to_remove) > 0):  # needed from ansible (list)
-    #         if(len(target_repos) > 0):# needed from ansible (list)
-    #             del_collaborators(g, target_repos, collaborators_to_remove, self.ORGANIZATION_NAME)
+        # Remove xuj1 from the repository to return it to the original state
 
-    #     output = get_collaborators(g, target_repos_copy)
-    #     collaborator_list = list()
-    #     for collaborator in output[target_repos_copy[0]]:
-    #         collaborator_list.append(collaborator['login'])
+        if(len(collaborators_to_remove) > 0):  # needed from ansible (list)
+            if(len(target_repos) > 0):# needed from ansible (list)
+                del_collaborators(g, target_repos, collaborators_to_remove, self.ORGANIZATION_NAME)
 
-    #     assert 'nk479217' not in collaborator_list
+        output = get_collaborators(g, target_repos_copy)
+        collaborator_list = list()
+        for collaborator in output[target_repos_copy[0]]:
+            collaborator_list.append(collaborator['login'])
 
-    # def test_pass_change_permissions_of_collaborator_enterprise_github(self):
-    #     g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
-    #     perms_check = "je652917"
-    #     target_repos = ['testing-repo-private']
+        assert self.USER not in collaborator_list
 
-    #     target_repos_copy = target_repos.copy()
-    #     if len(target_repos_copy):
-    #         for i in range(len(target_repos_copy)):
-    #             target_repos_copy[i] = self.ORGANIZATION_NAME + "/" + target_repos_copy[i]
+    def test_pass_change_permissions_of_collaborator_enterprise_github(self):
+        g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
+        perms_check = self.USER
+        target_repos = ['testing-repo-private']
 
-    #     change_collaborator_permissions(g, target_repos, perms_check, "triage", self.ORGANIZATION_NAME)
+        target_repos_copy = target_repos.copy()
+        if len(target_repos_copy):
+            for i in range(len(target_repos_copy)):
+                target_repos_copy[i] = self.ORGANIZATION_NAME + "/" + target_repos_copy[i]
 
-    #     output = get_collaborators(g, target_repos_copy)
+        change_collaborator_permissions(g, target_repos, perms_check, "push", self.ORGANIZATION_NAME)
 
-    #     collaborator_list = list()
-    #     for collaborator in output[target_repos_copy[0]]:
-    #         collaborator_list.append(collaborator['login'])
+        output = get_collaborators(g, target_repos_copy)
 
-    #     for i in range(0, len(collaborator_list)):
-    #         if output[target_repos_copy[0]][i]['login'] == 'je652917':
-    #             collaborator_permission = {'permissions': output[target_repos_copy[0]][i]['permissions']}
+        collaborator_list = list()
+        for collaborator in output[target_repos_copy[0]]:
+            collaborator_list.append(collaborator['login'])
 
-    #     assert 'triage=True' in str(collaborator_permission['permissions'])
+        for i in range(0, len(collaborator_list)):
+            if output[target_repos_copy[0]][i]['login'] == self.USER:
+                collaborator_permission = {'permissions': output[target_repos_copy[0]][i]['permissions']}
+
+        assert 'push=True' in str(collaborator_permission['permissions'])
+
+        change_collaborator_permissions(g, target_repos, perms_check, "pull", self.ORGANIZATION_NAME)
+
+        output = get_collaborators(g, target_repos_copy)
+
+        collaborator_list = list()
+        for collaborator in output[target_repos_copy[0]]:
+            collaborator_list.append(collaborator['login'])
+
+        for i in range(0, len(collaborator_list)):
+            if output[target_repos_copy[0]][i]['login'] == self.USER:
+                collaborator_permission = {'permissions': output[target_repos_copy[0]][i]['permissions']}
+
+        assert 'push=False' in str(collaborator_permission['permissions'])
+
+    def test_pass_check_for_unrealistic_role(self):
+        g = Github(self.API_KEY, base_url=self.ENTERPRISE_URL)
+        perms_check = self.USER
+        target_repos = ['testing-repo-internal']
+
+        target_repos_copy = target_repos.copy()
+        if len(target_repos_copy):
+            for i in range(len(target_repos_copy)):
+                target_repos_copy[i] = self.ORGANIZATION_NAME + "/" + target_repos_copy[i]
+
+        change_collaborator_permissions(g, target_repos, perms_check, "pizza", self.ORGANIZATION_NAME)
+
+        output = get_collaborators(g, target_repos_copy)
+
+        collaborator_list = list()
+        for collaborator in output[target_repos_copy[0]]:
+            collaborator_list.append(collaborator['login'])
+
+        for i in range(0, len(collaborator_list)):
+            if output[target_repos_copy[0]][i]['login'] == self.USER:
+                collaborator_permission = {'permissions': output[target_repos_copy[0]][i]['permissions']}
+
+        assert 'push=False' in str(collaborator_permission['permissions'])
+        assert 'pull=True' in str(collaborator_permission['permissions'])
 
     def test_pass_check_for_no_collaborators(self):
 
@@ -184,7 +228,8 @@ class TestCollaboratorRepositoryModule(unittest.TestCase):
         assert not collaborator_list
 
 if __name__ == '__main__':
-    if len(sys.argv) > 2:
+    if len(sys.argv) > 3:
+        TestCollaboratorRepositoryModule.USER = sys.argv.pop()
         TestCollaboratorRepositoryModule.ENTERPRISE_URL = sys.argv.pop()
         TestCollaboratorRepositoryModule.ORGANIZATION_NAME = sys.argv.pop()
         TestCollaboratorRepositoryModule.API_KEY = sys.argv.pop()
